set editor's background to the buffer panel
I found this was a way to apply white background (in the
default LAF, at least); the grey default caused a visible
boundary between the rightmost extent of the entered text
and the rest of the panel.

I would hope that also with other LAFs, this works sort of
and does not cause outlandish colors to be assigned.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4253 7869264c-0633-0410-bb99-d298f671b333