remove Actions interface from tabsetcontroller too
this is now a Listener interface in the notebook controller,
just like in the buffer controller case.

I thought for a moment using PropertyChangeSupport for this.
It would have been about the same amount of code though, and
a lot more fuzzy in terms of what a "property" is in this case.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4312 7869264c-0633-0410-bb99-d298f671b333