Notify user of a closed Resultset/more dynamic fetchsize
The problem is with the DB2 LUW driver. Unlike (as far as
the documentation goes) the Z/OS one, it does not support
the queryCloseImplicit property and closes exhausted ResultSets
after next(), so that no update it subsequently possible.

Every idea I could come up with to make this sort of dynamic
-- i. e. come up with heuristics to fetch with exact row size
-- is shit really and will confuse the user. Better have an
explicit input field for the fetch size.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4356 7869264c-0633-0410-bb99-d298f671b333