* restore saved-selection behavior * deselect
deselection is an experiment. I would have thought
setCaretPosition() was enough. However, I experienced strange
situations where a large block of text was replaced by an
insert in a way it would have been with an active selection.
I haven't been able to reproduce this, but as one try,
set selection start/end to caret position.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4479 7869264c-0633-0410-bb99-d298f671b333