try with custom PgUp/PgDown action on editor
I had really wanted to avoid that. However, not only did the
default action vary in its jump size a lot (I could probably find
out how to control that but that would not be easier), it also
wasn't able to scroll the viewport along with the movement;
it stopped at the visible viewport edge. Same story; I could have
found that out, I think, but all in all I figured implementing
a simple line-count-based variant was going to be quicker.

This one doesn't restore the caret offset from the line start.
I would like that; going to add it later.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4493 7869264c-0633-0410-bb99-d298f671b333