arrange from left with connections first
so that the default focus traversal order supports keyboard
selection of connections in a quick way.

This could be handled with a FocusTraversalPolicy too. But I
don't see much point. Left-aligned should be good as is.
Right-aligned was only to mimick SQLDeveloper, but I don't
think I really require it that way.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4512 7869264c-0633-0410-bb99-d298f671b333