Experiment: atomic undo for togglePrefix
I tried to read up on how to make a real 'edit'
(in the DocumentEvent sense) out of the toggle, but
I think it would have had to be a replace edit, leaving the
question again of tracking the selection change.

Now I use a proxy which delegates to UndoManager and can
invoke n-times in aggregate mode. This seems a little flimsy,
it'll definitely need a test drive.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@4522 7869264c-0633-0410-bb99-d298f671b333