* clean panel on split * scroll on result only * split resultless buffer too
it.1: not really needed but I think it will look cleaner on longer
running updates and such.

it.2: that should avoid a couple of irritiating scrolls on splitting
as we're populating result tables now without an actual fetch.

it.3: that was a regression. I wonder why I introduced that resultview
null check. I don't think it was mandated.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@6105 7869264c-0633-0410-bb99-d298f671b333