attempt showing connection in more locations, to avoid errors
I had a couple of attempts at this. Implementing a custom background
image seemed a bit cumbersome (and potentially ugly; not sure).
Using JLayeredPane to show some pattern on the right side or so
might have been an idea; but in the meantime I thought about the
possibility to implement kind of a "speech bubble" that follows
the cursor. I've tried that now with a simple JLabel. Let's see
if this is annoying or not.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@6549 7869264c-0633-0410-bb99-d298f671b333