don't wait for executor
not sure; I might reintroduce. I want to terminate; when is
the termination more complete: If a process spins down in the background
with the window already gone? Or if visible user feedback accompanies
the wait? How often would it even happen? And would it better to
reintroduce System.exit()? Now that I've given up on the daemon idea
for the tool processes.


git-svn-id: https://hq.steg0.eu/svn/trunk/tabletool@7383 7869264c-0633-0410-bb99-d298f671b333